api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
psycopg2-binary
alembic
pydantic>=2.0
pydantic-settings
python-multipart
pyyaml
ruamel.yaml
